requirements.txt
numpy>=1.20.3
torch>=1.13.1
transformers>=4.30.0
faiss-cpu>=1.7.4
tokenizers>=0.13.0
huggingface-hub>=0.14.0
langgraph>=0.2.0
cerebras-cloud-sdk>=1.0.0
Flask==3.1.0
pymongo>=4.6.0
python-dotenv>=1.0.0
flask-pymongo>=2.3.0
